python/django-djinn-backend/api/views.py
from rest_framework import viewsets
from djinn.models import Room
from api.serializers import RoomSerializer


class RoomViewSet(viewsets.ModelViewSet):
    queryset = Room.objects.all()
    serializer_class = RoomSerializer
