python/django-djinn-backend/requirements.txt
# Install required python packages using pip:
# pip install -r requirements.txt
Django==1.8.2
djangorestframework==3.1.3
